+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Last Changes
 
-- [#19](https://github.com/LaxarJS/ax-details-layer-widget/issues/19): prevent that an element in the background can get the focus by tabbing
+- [#19](https://github.com/LaxarJS/ax-details-layer-widget/issues/19): implemented alternative background scrolling prevention
+- [#17](https://github.com/LaxarJS/ax-details-layer-widget/issues/17): prevent that an element in the background can get the focus by tabbing
 
 
 ## v1.1.0